# packages/backend/src/generated/swagger.json
{
    "components": {
...
        "schemas": {
            "ApiErrorPayload": {
                "properties": {
                    "error": {
                        "properties": {
                            "data": {
                                "description": "Optional data containing details of the error"
                            },
                            "message": {
                                "type": "string",
                                "description": "A friendly message summarising the error"
                            },
                            "name": {
                                "type": "string",
                                "description": "Unique name for the type of error"
                            },
                            "statusCode": {
                                "type": "number",
                                "format": "integer",
                                "description": "HTTP status code"
                            }
                        },
                        "required": ["name", "statusCode"],
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["error"],
                        "nullable": false
                    }
                },
                "required": ["error", "status"],
                "type": "object",
                "description": "The Error object is returned from the api any time there is an error.\nThe message contains"
            },
...
            "ApiSuccessEmpty": {
                "properties": {
                    "results": {},
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["status"],
                "type": "object"
            },
...
            "Organization": {
                "properties": {
                    "defaultProjectUuid": {
                        "type": "string",
                        "description": "The project a user sees when they first log in to the organization"
                    },
                    "needsProject": {
                        "type": "boolean",
                        "description": "The organization needs a project if it doesn't have at least one project."
                    },
                    "chartColors": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "The default color palette for all projects in the organization"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the organization"
                    },
                    "organizationUuid": {
                        "type": "string",
                        "description": "The unique identifier of the organization",
                        "format": "uuid"
                    }
                },
                "required": ["name", "organizationUuid"],
                "type": "object",
                "description": "Details of a user's Organization"
            },
            "ApiOrganization": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/Organization"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "Pick_Organization.name_": {
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "The name of the organization"
                    }
                },
                "required": ["name"],
                "type": "object",
                "description": "From T, pick a set of properties whose keys are in the union K"
            },
            "CreateOrganization": {
                "$ref": "#/components/schemas/Pick_Organization.name_"
            },
            "Partial_Omit_Organization.organizationUuid-or-needsProject__": {
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "The name of the organization"
                    },
                    "chartColors": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "description": "The default color palette for all projects in the organization"
                    },
                    "defaultProjectUuid": {
                        "type": "string",
                        "description": "The project a user sees when they first log in to the organization"
                    }
                },
                "type": "object",
                "description": "Make all properties in T optional"
            },
            "UpdateOrganization": {
                "$ref": "#/components/schemas/Partial_Omit_Organization.organizationUuid-or-needsProject__"
            },
            "ProjectType": {
                "enum": ["DEFAULT", "PREVIEW"],
                "type": "string"
            },
...
            "OrganizationMemberRole": {
                "enum": [
                    "member",
                    "viewer",
                    "interactive_viewer",
                    "editor",
                    "developer",
                    "admin"
                ],
                "type": "string"
            },
...
            "DbtLocalProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.DBT"
                    },
                    "target": {
                        "type": "string"
                    },
                    "environment": {
                        "items": {
                            "$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
                        },
                        "type": "array"
                    },
                    "profiles_dir": {
                        "type": "string"
                    },
                    "project_dir": {
                        "type": "string"
                    }
                },
                "required": ["type"],
                "type": "object",
                "additionalProperties": false
            },
...
            "DbtCloudIDEProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.DBT_CLOUD_IDE"
                    },
                    "api_key": {
                        "type": "string"
                    },
                    "account_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    },
                    "environment_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    },
                    "project_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "number",
                                "format": "double"
                            }
                        ]
                    }
                },
                "required": [
                    "type",
                    "api_key",
                    "account_id",
                    "environment_id",
                    "project_id"
                ],
                "type": "object",
                "additionalProperties": false
            },
...
            "DbtNoneProjectConfig": {
                "properties": {
                    "type": {
                        "$ref": "#/components/schemas/DbtProjectType.NONE"
                    },
                    "target": {
                        "type": "string"
                    },
                    "environment": {
                        "items": {
                            "$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
                        },
                        "type": "array"
                    },
                    "hideRefreshButton": {
                        "type": "boolean"
                    }
                },
                "required": ["type"],
                "type": "object",
                "additionalProperties": false
            },
            "DbtProjectConfig": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/DbtLocalProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtCloudIDEProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtGithubProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtBitBucketProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtGitlabProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtAzureDevOpsProjectConfig"
                    },
                    {
                        "$ref": "#/components/schemas/DbtNoneProjectConfig"
                    }
                ]
            },
...
            "SnowflakeCredentials": {
                "$ref": "#/components/schemas/Omit_CreateSnowflakeCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "RedshiftCredentials": {
                "$ref": "#/components/schemas/Omit_CreateRedshiftCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "PostgresCredentials": {
                "$ref": "#/components/schemas/Omit_CreatePostgresCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "BigqueryCredentials": {
                "$ref": "#/components/schemas/Omit_CreateBigqueryCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "DatabricksCredentials": {
                "$ref": "#/components/schemas/Omit_CreateDatabricksCredentials.SensitiveCredentialsFieldNames_"
            },
...
            "TrinoCredentials": {
                "$ref": "#/components/schemas/Omit_CreateTrinoCredentials.SensitiveCredentialsFieldNames_"
            },
            "WarehouseCredentials": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/SnowflakeCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/RedshiftCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/PostgresCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/BigqueryCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/DatabricksCredentials"
                    },
                    {
                        "$ref": "#/components/schemas/TrinoCredentials"
                    }
                ]
            },
            "SupportedDbtVersions": {
                "enum": ["v1.4", "v1.5", "v1.6"],
                "type": "string"
            },
            "Project": {
                "properties": {
                    "dbtVersion": {
                        "$ref": "#/components/schemas/SupportedDbtVersions"
                    },
                    "copiedFromProjectUuid": {
                        "type": "string"
                    },
                    "pinnedListUuid": {
                        "type": "string"
                    },
                    "warehouseConnection": {
                        "$ref": "#/components/schemas/WarehouseCredentials"
                    },
                    "dbtConnection": {
                        "$ref": "#/components/schemas/DbtProjectConfig"
                    },
                    "type": {
                        "$ref": "#/components/schemas/ProjectType"
                    },
                    "name": {
                        "type": "string"
                    },
                    "projectUuid": {
                        "type": "string"
                    },
                    "organizationUuid": {
                        "type": "string"
                    }
                },
                "required": [
                    "dbtVersion",
                    "dbtConnection",
                    "type",
                    "name",
                    "projectUuid",
                    "organizationUuid"
                ],
                "type": "object"
            },
            "ApiProjectResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/Project"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
...
            "LightdashUser": {
                "properties": {
                    "userUuid": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "organizationUuid": {
                        "type": "string"
                    },
                    "organizationName": {
                        "type": "string"
                    },
                    "organizationCreatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "isTrackingAnonymized": {
                        "type": "boolean"
                    },
                    "isMarketingOptedIn": {
                        "type": "boolean"
                    },
                    "isSetupComplete": {
                        "type": "boolean"
                    },
                    "role": {
                        "$ref": "#/components/schemas/OrganizationMemberRole"
                    },
                    "isActive": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "userUuid",
                    "firstName",
                    "lastName",
                    "isTrackingAnonymized",
                    "isMarketingOptedIn",
                    "isSetupComplete",
                    "isActive"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ApiGetAuthenticatedUserResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/LightdashUser"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object",
                "description": "Shows the authenticated user"
            },
            "ApiRegisterUserResponse": {
                "properties": {
                    "results": {
                        "$ref": "#/components/schemas/LightdashUser"
                    },
                    "status": {
                        "type": "string",
                        "enum": ["ok"],
                        "nullable": false
                    }
                },
                "required": ["results", "status"],
                "type": "object"
            },
            "ActivateUser": {
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    }
                },
                "required": ["password", "lastName", "firstName"],
                "type": "object"
            },
            "ActivateUserWithInviteCode": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ActivateUser"
                    },
                    {
                        "properties": {
                            "inviteCode": {
                                "type": "string"
                            },
                            "ownIdData": {
                                "type": "string"
                            }
                        },
                        "required": ["inviteCode"],
                        "type": "object"
                    }
                ]
            },
            "Email": {
                "type": "string",
                "description": "Email",
                "pattern": "^(([^<>()[\\]\\\\.,;:\\s@\"]+(\\.[^<>()[\\]\\\\.,;:\\s@\"]+)*)|(\".+\"))@((\\[[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\.[0-9]{1,3}\\])|(([a-zA-Z\\-0-9]+\\.)+[a-zA-Z]{2,}))$"
            },
            "CreateUserArgs": {
                "properties": {
                    "ownIdData": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "email": {
                        "$ref": "#/components/schemas/Email"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    }
                },
                "required": ["password", "email", "lastName", "firstName"],
                "type": "object"
            },
            "RegisterOrActivateUser": {
                "anyOf": [
                    {
                        "$ref": "#/components/schemas/ActivateUserWithInviteCode"
                    },
                    {
                        "$ref": "#/components/schemas/CreateUserArgs"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "session_cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "connect.sid"
            },
            "api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Value should be 'ApiKey <your key>'"
            }
        }
    },
    "info": {
        "title": "Lightdash API",
        "version": "0.765.1",
        "description": "Open API documentation for all public Lightdash API endpoints",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "Lightdash Support",
            "email": "support@lightdash.com",
            "url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/org": {
            "get": {
                "operationId": "GetMyOrganization",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiOrganization"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get the current user's organization",
                "tags": ["Organizations"],
                "security": [],
                "parameters": []
            },
            "put": {
                "operationId": "CreateOrganization",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessEmpty"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Creates a new organization, the current user becomes the Admin of the new organization.\nThis is only available to users that are not already in an organization.",
                "tags": ["Organizations"],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "description": "the new organization settings",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrganization",
                                "description": "the new organization settings"
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "UpdateMyOrganization",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessEmpty"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Update the current user's organization",
                "tags": ["Organizations"],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "description": "the new organization settings",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrganization",
                                "description": "the new organization settings"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/projects/{projectUuid}": {
            "get": {
                "operationId": "GetProject",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiProjectResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get a project of an organiztion",
                "tags": ["Projects"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "projectUuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/v1/user": {
            "get": {
                "operationId": "GetAuthenticatedUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiGetAuthenticatedUserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Get authenticated user",
                "tags": ["My Account"],
                "security": [],
                "parameters": []
            },
            "post": {
                "operationId": "RegisterUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiRegisterUserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorPayload"
                                }
                            }
                        }
                    }
                },
                "description": "Register user",
                "tags": ["My Account"],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterOrActivateUser"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
}